CHANGE APPROVERS LIST IN OWNERS FILE
According to best practices in the open source community
only the people that are actively reviewing code should be
promoted to approver.

Ideally people should only become a member of a project if
they are contributing regularly. And only become reviewers
when they are familiar with the project and code base.

As an exception our team is accepted in the reviewers list.
But only approvers have merge permissions.