CLEAN TYPES FOR PRIMARY INTERFACES
The primary interface in a pod is not owned by this operator/controller.
It's a bad practice making changes to objects or resources that are not owned by the operator.
Configurations may be impacted by changes in the primary CNI plugin.
Therefore, the MTU story is being droppped for the purposes of better alignment
with community premises declared in other documents such as the Multus
design documentation around a Kubernetes Network CRD de facto standard.

That can be found here:
https://docs.google.com/document/d/1Ny03h6IDVy_e_vmElOqR7UdTPAG_RNydhVE1Kx54kFQ/edit#heading=h.hylsbqoj5fxd